Moved config schema and config type to src
