+++ apps/pxweb2/public/theme/variables.css
@@ -1,6 +1,6 @@
 /**
  * Do not edit directly
- * Generated on Fri, 01 Mar 2024 13:12:05 GMT
+ * Generated on Wed, 17 Apr 2024 13:00:42 GMT
  */
 
 :root {
