+++ apps/pxweb2/src/app/util/config/configType.ts
