+++ apps/pxweb2/src/app/util/config/getConfig.ts
@@ -1,4 +1,4 @@
-import { Config } from '../../../public/config/configType';
+import { Config } from '../../../../public/config/configType';
 export const getConfig = () => {
   return (window as any).PxWeb2Config as Config;
 };
